backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
channels==4.3.1
channels-redis==4.3.0
psycopg[binary]==3.2.10
stripe==13.0.0
google-generativeai==0.8.5
